requirements.txt
# This requirements file lists all third-party dependencies for this project.
#
# Run 'pip install -r requirements.txt -t lib/' to install these dependencies
# in `lib/` subdirectory.
#
# Note: The `lib` directory is added to `sys.path` by `appengine_config.py`.
Flask==0.11
simplejson==3.8.2
passlib==1.6.5
blinker==1.4
transitions==0.4.0
marshmallow==2.9.1

# db cassandra
cassandra-driver==3.6.0
